feat: Todo list and hr
