+++ starter-kits/nextjs-starter-kit/components/TextEditor/Toolbar.tsx
@@ -1,8 +1,9 @@
 import { Editor } from "@tiptap/react";
+import { ToolbarInlineAdvanced } from "./TextInlineAdvanced";
 import { ToolbarAlignment } from "./ToolbarAlignment";
 import { ToolbarBlock } from "./ToolbarBlock";
 import { ToolbarCommands } from "./ToolbarCommands";
-import { Headings } from "./ToolbarHeadings";
+import { ToolbarHeadings } from "./ToolbarHeadings";
 import { ToolbarInline } from "./ToolbarInline";
 import { ToolbarMedia } from "./ToolbarMedia";
 import styles from "./Toolbar.module.css";
@@ -16,9 +17,11 @@ export function Toolbar({ editor }: Props) {
     <div className={styles.toolbar}>
       <ToolbarCommands editor={editor} />
       <div className={styles.toolbarSeparator} />
-      <Headings editor={editor} />
+      <ToolbarHeadings editor={editor} />
       <div className={styles.toolbarSeparator} />
       <ToolbarInline editor={editor} />
+      <div className={styles.toolbarSeparator} />
+      <ToolbarInlineAdvanced editor={editor} />
       <div className={styles.toolbarSeparator} />
       <ToolbarAlignment editor={editor} />
       <div className={styles.toolbarSeparator} />
