+++ starter-kits/nextjs-starter-kit/components/TextEditor/ToolbarInline.tsx
@@ -1,13 +1,8 @@
 import { Editor } from "@tiptap/react";
-import { useState } from "react";
-import { CrossIcon, LinkIcon } from "../../icons";
 import { BoldIcon } from "../../icons/Bold";
-import { CodeIcon } from "../../icons/Code";
 import { ItalicIcon } from "../../icons/Italic";
 import { StrikethroughIcon } from "../../icons/Strikethrough";
 import { Button } from "../../primitives/Button";
-import { Input } from "../../primitives/Input";
-import { Popover } from "../../primitives/Popover";
 import styles from "./Toolbar.module.css";
 
 type Props = {
@@ -15,13 +10,6 @@ type Props = {
 };
 
 export function ToolbarInline({ editor }: Props) {
-  function toggleLink(link: string) {
-    console.log(link);
-    editor.chain().focus().toggleLink({ href: link }).run();
-  }
-
-  // console.log(editor.getAttributes("link").href);
-
   return (
     <>
       <Button
@@ -56,81 +44,6 @@ export function ToolbarInline({ editor }: Props) {
       >
         <StrikethroughIcon />
       </Button>
-
-      <Button
-        variant="subtle"
-        className={styles.toolbarButton}
-        onClick={() => editor.chain().focus().toggleCode().run()}
-        disabled={!editor.can().chain().focus().toggleCode().run()}
-        data-active={editor.isActive("code") ? "is-active" : undefined}
-        aria-label="Strikethrough"
-      >
-        <CodeIcon />
-      </Button>
-
-      <Popover
-        content={
-          <LinkPopover
-            onSubmit={toggleLink}
-            onRemoveLink={toggleLink}
-            showRemove={editor.getAttributes("link").href}
-          />
-        }
-      >
-        <Button
-          variant="subtle"
-          className={styles.toolbarButton}
-          disabled={!editor.can().chain().focus().setLink({ href: "" }).run()}
-          data-active={editor.isActive("link") ? "is-active" : undefined}
-          aria-label="Link"
-        >
-          <LinkIcon style={{ width: "17px" }} />
-        </Button>
-      </Popover>
     </>
-  );
-}
-
-type LinkPopoverProps = {
-  onSubmit: (url: string) => void;
-  onRemoveLink: (url: string) => void;
-  showRemove: boolean;
-};
-
-function LinkPopover({ onSubmit, onRemoveLink, showRemove }: LinkPopoverProps) {
-  const [value, setValue] = useState("");
-
-  return (
-    <form
-      className={styles.toolbarPopover}
-      onSubmit={(e) => {
-        e.preventDefault();
-        onSubmit(value);
-      }}
-    >
-      <label className={styles.toolbarPopoverLabel} htmlFor="">
-        Add link to selected text
-      </label>
-      <div className={styles.toolbarPopoverBar}>
-        <Input
-          className={styles.toolbarPopoverInput}
-          value={value}
-          onChange={(e) => setValue(e.target.value)}
-        />
-        {showRemove ? (
-          <Button
-            variant="secondary"
-            onClick={(e) => {
-              e.stopPropagation();
-              onRemoveLink(value);
-            }}
-            aria-label="Remove link"
-          >
-            <CrossIcon />
-          </Button>
-        ) : null}
-        <Button className={styles.toolbarPopoverButton}>Add link</Button>
-      </div>
-    </form>
   );
 }
