+++ starter-kits/nextjs-starter-kit/package.json
@@ -22,9 +22,12 @@
     "@tiptap/extension-character-count": "^2.0.4",
     "@tiptap/extension-collaboration": "^2.0.4",
     "@tiptap/extension-collaboration-cursor": "^2.0.4",
+    "@tiptap/extension-highlight": "^2.0.4",
     "@tiptap/extension-image": "^2.0.4",
     "@tiptap/extension-link": "^2.0.4",
     "@tiptap/extension-placeholder": "^2.0.4",
+    "@tiptap/extension-task-item": "^2.0.4",
+    "@tiptap/extension-task-list": "^2.0.4",
     "@tiptap/extension-text-align": "^2.0.4",
     "@tiptap/extension-typography": "^2.0.4",
     "@tiptap/extension-youtube": "^2.0.4",
